Added more possibilities for opening files with different access levels.
Added the option to supply the owner password since in some cases this
is required to stamp the PDF. Also enabled an override that prevents
some PDFs with the appropriate permissions from being stamped. Hopefully
this method also allows opening of encrypted PDFs (I have no motivation
to test this right now).